抽取出DAO，专门处理与数据库相关的连接和增删改查操作
